requirements.txt
asgiref==3.7.2
Django==4.2.3
django-rest-auth==0.9.5
djangorestframework==3.14.0
drf-yasg==1.21.6
inflection==0.5.1
packaging==23.1
pytz==2023.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
